pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iot-data-bridge"
version = "0.1.0"
description = "IoT Data Bridge - External data to Device mapping middleware"
authors = [{name = "IoT Data Bridge Team"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiomqtt>=2.0.0",
    "signalrcore>=0.9.3",
    "pydantic>=2.5.0",
    "structlog>=23.2.0",
    "pyyaml>=6.0.1",
    "asyncio-mqtt>=0.16.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
